+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Main from '@/pages/Main';
-import Login from '@/pages/Login';
+import Main from './pages/Main';
+import Login from './pages/Login';
 
 function App() {
   return (
